Smaller bug fixes

+++ generate_savanna.m
@@ -9,5 +9,5 @@ function savanna = generate_savanna(pine_ratio, hw_ratio)
     hw_end_i = uint64(2500*(pine_ratio+hw_ratio));
     savanna(vec(1:pine_end_i)) = uint8(rand(pine_end_i,1)*get_repr('young')) + 1;
     savanna(vec(pine_end_i+1:hw_end_i)) = ...
-        uint64(rand(hw_end_i-pine_end_i,1)*get_repr('young')) + get_repr('young_hw');
+        uint8(rand(hw_end_i-pine_end_i,1)*get_repr('young')) + get_repr('young_hw');
 end
